+++ tests/datetime/test_from_format.py
@@ -37,6 +37,12 @@ def test_from_format_with_escaped_elements():
     assert "+00:00" == d.timezone_name
 
 
+def test_from_format_with_escaped_elements_valid_tokens():
+    d = pendulum.from_format("1975-05-21T22:32:11.123Z", "YYYY-MM-DD[T]HH:mm:ss.SSS[Z]")
+    assert_datetime(d, 1975, 5, 21, 22, 32, 11)
+    assert "UTC" == d.timezone_name
+
+
 def test_from_format_with_millis():
     d = pendulum.from_format("1975-05-21 22:32:11.123456", "YYYY-MM-DD HH:mm:ss.SSSSSS")
     assert_datetime(d, 1975, 5, 21, 22, 32, 11, 123456)