+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fixed missing `x` token for string formatting.
 - Fixed reading timezone files.
 - Added support for parsing padded 2-digit days of the month with `from_format()`
+- Fixed `from_format()` trying to parse escaped tokens.
 
 
 ## [2.0.3] - 2018-07-30